backend/requirements.txt
# ===== BEGIN requirements.txt =====
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic-settings==2.1.0
python-dotenv==1.0.0
openai==1.3.8
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
PyPDF2==3.0.1
python-docx==1.1.0
redis==5.0.1
rq==1.15.1
httpx==0.25.2
python-dateutil==2.8.2
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
ruff==0.1.6
# ===== END requirements.txt =====
